backend/requirements.txt
flask
psycopg2-binary 
flask_sqlalchemy
python-dotenv